\name{pmt-package}
\alias{pmt-package}
\alias{pmt}
\docType{package}
\title{
\packageTitle{pmt}
Poisson Margin Test
}
\description{
\packageDescription{pmt}
Computes p-values using the Poission Margin Test for two vectors of counts arising from NGS data.
}
\details{
\packageDESCRIPTION{pmt}
\packageIndices{pmt}
}
\author{
\packageAuthor{pmt}

Maintainer: \packageMaintainer{pmt}
}
\references{
Kowalczyk, A., Bedő, J., Conway, T., Beresford-Smith, B., 2011. The Poisson margin test for normalization-free significance analysis of NGS data. Journal of Computational Biology 18, 391–400. doi:10.1089/cmb.2010.0272
}
\keyword{NGS}
\keyword{Poisson}
